Polishing of the code.
