+++ app.json
@@ -5,6 +5,7 @@
     "name": "WashroomApp",
     "slug": "WashroomApp",
     "version": "1.0.0",
+    "icon": "./assets/appIcon.png",
     "assetBundlePatterns": [
       "**/*"
     ]
