+++ components/globalStyle.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+export const globalStyles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: '#fff',
+      //   alignItems: 'center',
+        justifyContent: 'center',
+        flexDirection: 'column',
+      },
+      containerAcross: {
+        flex: 1, 
+        flexDirection: 'row',
+        justifyContent: 'center',
+      },
+
+
+})
