+++ screens/watcher.js
@@ -1,0 +1,27 @@
+import React, { useState, useEffect } from 'react';
+import { ImageBackground, StyleSheet, View, Text, TouchableOpacity, TouchableWithoutFeedback, Keyboard, 
+    TextInput, Alert } from 'react-native';
+import Card from '../components/card';
+import { MaterialIcons } from '@expo/vector-icons';
+import { FlatList } from 'react-native-gesture-handler';
+
+export default function Watcher({ navigation }) {
+
+    return (
+        <View style={styles.container}>
+
+        </View>
+    );
+}
+
+const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: '#fff',
+    //   alignItems: 'center',
+      justifyContent: 'center',
+      flexDirection: 'column',
+      
+    },
+  });
+
